# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an Inventario daemon.",
        "title": "Inventario API",
        "contact": {
            "name": "Inventario Support",
            "url": "https://github.com/denisvmedia/inventario/issues",
            "email": "ask@artprima.cz"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/g/{groupSlug}/settings": {
            "get": {
                "description": "get current settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get current settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group slug",
                        "name": "groupSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SettingsObject"
                        }
                    }
                }
            },
            "put": {
                "description": "update entire settings object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group slug",
                        "name": "groupSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Settings object with documented snake_case field names",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apiserver.SettingsUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SettingsObject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/g/{groupSlug}/settings/{field}": {
            "patch": {
                "description": "update a specific setting field.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Patch setting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group slug",
                        "name": "groupSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Setting field path (e.g., uiconfig.theme)",
                        "name": "field",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Setting value envelope with required value.",
                        "name": "value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apiserver.PatchSettingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SettingsObject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apiserver.PatchSettingRequest": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Value is the setting value to apply and is required when using the object envelope."
                }
            }
        },
...
        "apiserver.SettingsUpdateRequest": {
            "type": "object",
            "properties": {
                "default_date_format": {
                    "description": "DefaultDateFormat is the uiconfig.default_date_format value accepted by PUT /settings.",
                    "type": "string"
                },
                "show_debug_info": {
                    "description": "ShowDebugInfo is the uiconfig.show_debug_info value accepted by PUT /settings.",
                    "type": "boolean"
                },
                "theme": {
                    "description": "Theme is the uiconfig.theme value accepted by PUT /settings.",
                    "type": "string"
                }
            }
        },
...
        "models.SettingsObject": {
            "type": "object",
            "properties": {
                "defaultDateFormat": {
                    "type": "string"
                },
                "showDebugInfo": {
                    "type": "boolean"
                },
                "theme": {
                    "type": "string"
                }
            }
        },
...
    }
}